requirements.txt
Werkzeug==1.0.1
clarifai==2.6.2
Flask==1.1.2
requests==2.22.0
python-dotenv==0.14.0
gunicorn==19.5.0
gevent==20.6.2
sqlalchemy==1.3.18
clarifai-grpc==7.4.0 
googleapis-common-protos==1.53.0 
grpcio==1.37.1 
protobuf==3.15.8
requests==2.25.1